完成秒杀显示，但未完成秒杀详情页

+++ pyg-dao/src/main/java/top/ingxx/mapper/TbSeckillGoodsMapper.java
@@ -7,7 +7,7 @@ import top.ingxx.pojo.TbSeckillGoodsExample;
 import java.util.List;
 
 public interface TbSeckillGoodsMapper {
-    int countByExample(TbSeckillGoodsExample example);
+    long countByExample(TbSeckillGoodsExample example);
 
     int deleteByExample(TbSeckillGoodsExample example);
